Delete dead job_queue module and the legacy PPTX upload flow
job_queue.py was never imported at runtime - the backend runs its own
TranslationJob system - so it goes, along with its tests; the dashboard
test moves to its own file since translation_metrics IS live (the plan
note calling it dead was wrong, now corrected). The unreferenced
handle_upload PPTX flow goes too.
